Métodos getByID, Delete, Update, getAll e Create de Cells

+++ src/main/java/br/fiocruz/procc/acbmservice/repository/CellRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CellRepository extends CrudRepository<Cell, Integer> {
+public interface CellRepository extends CrudRepository<Cell, Long> {
 }
